Correct URL for Mura
The previous github url no longer works, as Mura is no longer open source. I have offered the url for the company, though it's not clear if it's downloadable or even based on CFML. I didn't want to presume that these points, alone, might justify REMOVING it. I'll leave that for you to consider.